ai/requirements.txt
timm
tqdm
torch
transformers
tf-keras
pillow
torchvision
huggingface_hub
open_clip_torch
opencv-python
einops
celery
redis
requests
python-dotenv
